Reset with base auth resolvers and iron-session

+++ src/utils/sessions.ts
@@ -1,8 +1,14 @@
-import { GetServerSidePropsContext } from 'next';
-import { IncomingMessage } from 'http';
-import { applySession, SessionOptions } from 'next-iron-session';
-import { Session, User } from '@prisma/client';
+import { addSeconds, differenceInSeconds } from 'date-fns';
 import { db } from './prisma';
+import { getIronSession, IronSession, IronSessionOptions } from 'iron-session';
+import { IncomingMessage, ServerResponse } from 'http';
+import { Session, User } from '@prisma/client';
+
+declare module 'iron-session' {
+  interface IronSessionData {
+    sessionID?: string | null;
+  }
+}
 
 if (!process.env.COOKIE_SECRET) {
   console.warn(
@@ -14,14 +20,12 @@ if (!process.env.COOKIE_SECRET) {
 // The session will automatically renew when there's < 25% of it validity period
 const SESSION_TTL = 1 * 24 * 3600 * 1000;
 
-const IRON_SESSION_ID_KEY = 'sessionID';
-
 interface ReqWithSession extends IncomingMessage {
-  session: import('next-iron-session').Session;
+  session: import('iron-session').IronSession;
 }
 
-const sessionOptions: SessionOptions = {
-  password: [{ id: 1, password: process.env.COOKIE_SECRET as string }],
+const sessionOptions: IronSessionOptions = {
+  password: { 1: process.env.COOKIE_SECRET as string },
   cookieName: 'session.info',
   ttl: SESSION_TTL,
   cookieOptions: {
@@ -32,7 +36,7 @@ const sessionOptions: SessionOptions = {
   }
 };
 
-export async function createSession(req: IncomingMessage, user: User) {
+export async function createSession(ironSession: IronSession, user: User) {
   const session = await db.session.create({
     data: {
       userId: user.id,
@@ -40,48 +44,60 @@ export async function createSession(req: IncomingMessage, user: User) {
     }
   });
 
-  const sessionID = session.id;
+  ironSession.sessionID = session.id;
 
-  const reqWithSession = req as unknown as ReqWithSession;
-  reqWithSession.session.set(IRON_SESSION_ID_KEY, sessionID);
-
-  await reqWithSession.session.save();
+  await ironSession.save();
 
   return session;
 }
 
-export async function removeSession(req: IncomingMessage, session: Session) {
-  const reqWithSession = req as unknown as ReqWithSession;
+export async function removeSession(
+  ironSession: IronSession,
+  session?: Session | null
+) {
+  ironSession.destroy();
 
-  await db.session.delete({ where: { id: session.id } });
+  await ironSession.save();
 
-  reqWithSession.session.destroy();
+  if (session) {
+    await db.session.delete({ where: { id: session.id } });
+  }
 }
 
-const sessionCache = new WeakMap<IncomingMessage, Session | null>();
-export async function resolveSession({
-  req,
-  res
-}: Pick<GetServerSidePropsContext, 'req' | 'res'>) {
-  if (sessionCache.has(req)) {
-    return sessionCache.get(req);
+interface CachedSession {
+  session: Session | null;
+  ironSession: IronSession;
+}
+
+const sessionCache = new WeakMap<IncomingMessage, CachedSession>();
+
+export async function resolveSession(
+  req: IncomingMessage,
+  res: ServerResponse
+): Promise<CachedSession> {
+  const cacheSession = sessionCache.get(req);
+
+  if (cacheSession) {
+    return cacheSession;
   }
 
-  await applySession(req, res, sessionOptions);
+  const ironSession = await getIronSession(req, res, sessionOptions);
+
+  const sessionID = ironSession.sessionID;
 
   let session: Session | null = null;
 
-  const reqWithSession = req as unknown as ReqWithSession;
-  const sessionID = reqWithSession.session.get(IRON_SESSION_ID_KEY);
-
   if (sessionID) {
     session = await db.session.findFirst({
-      where: { id: sessionID, expiresAt: { gte: new Date() } }
+      where: { id: sessionID, expiresAt: { gte: new Date() } },
+      include: {
+        user: true
+      }
     });
 
     if (session) {
       const shouldRefreshSession =
-        Date.now() - session.expiresAt.getTime() < 0.75 * SESSION_TTL;
+        differenceInSeconds(session.expiresAt, new Date()) < 0.75 * SESSION_TTL;
 
       if (shouldRefreshSession) {
         await db.session.update({
@@ -89,16 +105,20 @@ export async function resolveSession({
             id: session.id
           },
           data: {
-            expiresAt: new Date(Date.now() + SESSION_TTL)
+            expiresAt: addSeconds(new Date(), SESSION_TTL)
           }
         });
 
-        await reqWithSession.session.save();
+        await ironSession.save();
       }
+    } else {
+      ironSession.destroy();
+
+      await ironSession.save();
     }
   }
 
-  sessionCache.set(req, session);
+  sessionCache.set(req, { session, ironSession });
 
-  return session;
+  return { session, ironSession };
 }
